collecting stats from WIG initial
