+++ src/OrleansRuntime/Scheduler/PoliciedScheduler/PriorityBasedTaskScheduler.cs
@@ -6,6 +6,7 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 using Orleans.Runtime.Configuration;
+using Orleans.Runtime.Scheduler.Utility;
 
 namespace Orleans.Runtime.Scheduler.PoliciedScheduler
 {
@@ -27,6 +28,10 @@ namespace Orleans.Runtime.Scheduler.PoliciedScheduler
 
         private readonly Dictionary<short, Tuple<ulong, HashSet<ulong>>> tenants;
         private readonly Dictionary<short, long> timeLimitsOnTenants;
+        private readonly Dictionary<WorkItemGroup, FixedSizedQueue<double>> tenantStatCounters;
+        private const int MaximumStatCounterSize = 100;
+        // TODO: FIX LATER
+        private int statCollectionCounter = 100;
         #endregion
         #endregion
 
@@ -63,8 +68,10 @@ namespace Orleans.Runtime.Scheduler.PoliciedScheduler
                     config.TurnWarningLengthThreshold, config.EnableWorkerThreadInjection, config.LimitManager.GetLimit(LimitNames.LIMIT_MAX_PENDING_ITEMS),
                     performanceMetrics)
         {
+            // Tenancies
             tenants = new Dictionary<short, Tuple<ulong, HashSet<ulong>>>();
             timeLimitsOnTenants = new Dictionary<short, long>();
+            tenantStatCounters = new Dictionary<WorkItemGroup, FixedSizedQueue<double>>();
         }
 
         private PriorityBasedTaskScheduler(int maxActiveThreads, TimeSpan delayWarningThreshold, TimeSpan activationSchedulingQuantum,
@@ -175,6 +182,14 @@ namespace Orleans.Runtime.Scheduler.PoliciedScheduler
                 var t = TaskSchedulerUtils.WrapWorkItemWithPriorityAsTask(workItem, priorityContext, workItemGroup.TaskRunner);
                 t.Start(workItemGroup.TaskRunner);
             }
+
+            //TODO: FIX LATER
+            if (--statCollectionCounter <= 0)
+            {
+                statCollectionCounter = 100;
+                foreach (var kv in tenantStatCounters) tenantStatCounters[kv.Key].Enqueue(kv.Key.CollectStats());
+                logger.Info($"Priniting execution times in ticks: {string.Join("********************",tenantStatCounters.Select(x=>x.Key.ToString()+':' + String.Join(",", x.Value)))}");
+            }
         }
 
         public void QueueControllerWorkItem(IWorkItem workItem, ISchedulingContext context)
@@ -197,16 +212,28 @@ namespace Orleans.Runtime.Scheduler.PoliciedScheduler
             var controllerContext = ((InvokeWorkItem)workItem).ControllerContext;
 
             ulong controllerId = ((InvokeWorkItem)workItem).SourceActivation.Grain.Key.N1;
+            var schedulingContext = context as SchedulingContext;
             if (tenants.ContainsKey(controllerContext.AppId))
             {
-                var schedulingContext = context as SchedulingContext;
                 tenants[controllerContext.AppId].Item2.Add(schedulingContext.Activation.Grain.Key.N1);
             }
             else
             {
+                // Initialize entries in *ALL* per-dataflow maps
                 tenants.Add(controllerContext.AppId, new Tuple<ulong, HashSet<ulong>>(controllerId, new HashSet<ulong>()));
                 timeLimitsOnTenants.Add(controllerContext.AppId, controllerContext.Time);
+
+                tenants[controllerContext.AppId].Item2.Add(schedulingContext.Activation.Grain.Key.N1);
             }
+            var wig = GetWorkItemGroup(schedulingContext);
+            if (wig==null)
+            {
+                var error = string.Format(
+                    "WorkItem {0} on context {1} does not match a work item group", workItem, context);
+                logger.Error(ErrorCode.SchedulerQueueWorkItemWrongCall, error);
+                throw new InvalidOperationException(error);
+            }
+            if (!tenantStatCounters.ContainsKey(wig)) tenantStatCounters.Add(wig, new FixedSizedQueue<double>(MaximumStatCounterSize));
             
             QueueWorkItem(workItem, context);
         }
